# docs/openapi-esa.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Art Data API",
    "description": "ESA API schema validation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.wx.2000gallery.art",
      "description": "production"
    }
  ],
  "components": {
    "schemas": {
      "AuthRegister": {
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthLogin": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "WxRefreshTokenBody": {
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "CartAddItem": {
        "type": "object",
        "required": [
          "type",
          "quantity"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "right_id": {
            "type": "integer",
            "format": "int32"
          },
          "digital_artwork_id": {
            "type": "integer",
            "format": "int32"
          },
          "artwork_id": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "FavoriteAdd": {
        "type": "object",
        "required": [
          "item_type",
          "item_id"
        ],
        "properties": {
          "item_type": {
            "type": "string"
          },
          "item_id": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "postAuthRegister",
        "summary": "POST /api/auth/register",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegister"
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "postAuthLogin",
        "summary": "POST /api/auth/login",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLogin"
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "postAuthRefresh",
        "summary": "POST /api/auth/refresh",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WxRefreshTokenBody"
              }
            }
          }
        }
      }
    },
...
    "/api/digital-identity/purchases/{user_id}": {
      "get": {
        "operationId": "getDigitalIdentityPurchases",
        "summary": "GET /api/digital-identity/purchases/{user_id}",
        "tags": [
          "digital-identity"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
    "/api/cart": {
      "get": {
        "operationId": "get_cart",
        "summary": "GET /api/cart",
        "tags": [
          "cart"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      },
      "post": {
        "operationId": "post_cart",
        "summary": "POST /api/cart",
        "tags": [
          "cart"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartAddItem"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_cart",
        "summary": "DELETE /api/cart",
        "tags": [
          "cart"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
...
    "/api/favorites": {
      "get": {
        "operationId": "get_favorites",
        "summary": "GET /api/favorites",
        "tags": [
          "favorites"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      },
      "post": {
        "operationId": "post_favorites",
        "summary": "POST /api/favorites",
        "tags": [
          "favorites"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteAdd"
              }
            }
          }
        }
      }
    },
...
    "/api/wx/pay/digital-identity/purchases/{user_id}": {
      "get": {
        "operationId": "get_wx_pay_digital_identity_purchases_user_id",
        "summary": "GET /api/wx/pay/digital-identity/purchases/{user_id}",
        "tags": [
          "wechat-pay"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "get_search",
        "summary": "GET /api/search",
        "tags": [
          "search"
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}